# Optimize btree query