Requires CTE and thus Postgres 8.4. Here is a naïve comparison :

``` console
$ time psql btree_bloat-8.2-8.3.sql > old.out          
                                                                                                                                                                                                          
real    0m2,087s
user    0m0,035s
sys     0m0,011s
$ time psql btree_bloat.sql > new.out

real    0m0,107s
user    0m0,043s
sys     0m0,000s
$ diff -u old.out new.out 
$
```

cc @madtibo