@madtibo 2020-08-13
|  PostgreSQL version             |    9.5      |    9.6      |    10     |     11    |     12    |
|---------------------------------|-------------|-------------|-----------|-----------|-----------|
| btree/btree_bloat-superuser.sql |   30,836 ms | 31,127 ms   | 26,710 ms | 25,374 ms | 25,633 ms |
| btree/btree_bloat.sql           |   48,019 ms | 48,775 ms   | 61,816 ms | 59,598 ms | 82,369 ms |
| btree/btree_bloat_old.sql       | 3845,079 ms | 3828,218 ms | 79,153 ms | 74,919 ms | 80,893 ms |