@madtibo 2020-08-14
La contre performance en 12 vient du fait que le CTE est traité comme un sous-select. On retombe donc dans un sous select non optimal. Les perf sont meilleures avec un WITH avec le mot clé MATERIALIZED.

|  PostgreSQL version             |  12    |
|---------------------------------|--------|
| btree/btree_bloat-superuser.sql | 25,695 ms |
| btree/btree_bloat-materialized.sql       | 61,488 ms |
| btree/btree_bloat.sql           | 82,015 ms |
| btree/btree_bloat-8.2-8.3.sql   | 82,156 ms |

Le problème, c'est que le mot clé MATERIALIZED n'arrive que en version 12...
Est-ce qu'on crée une nouvelle version de la requête ?