@ioguix 2020-08-17
Hi,

I might be wrong, but it seems related to #9.

I don't understand why you compare with the 8.2-8.3 query as it seem your problem is on a recent release. 

Moreover, maybe you should share the data you are testing on so I can reproduce.

Based on https://github.com/ioguix/pgsql-bloat-estimation/pull/16#issuecomment-673546794, the running time seem pretty decent to me.

I prefer avoiding the CTE as long as possible as it brings its own disadvantages.