@madtibo 2020-08-24
It is indeed the same problem as raised by julmon.

The query was split for 8.2 and 8.3 versions since CTE did not exist at the time.

The running time with the current query is really bad for versions 9.6 and before.
The CTE is important to include for these versions.